+++ ipc_sun_sync/main.py
@@ -4,6 +4,7 @@ import traceback
 import pytz
 
 from . import __version__
+from .constants import IPC, ConfigIPC, ConfigMethod
 from .dahua_cgi import DahuaCgi
 from .dahua_rpc import DahuaRpc
 from .parser import parse_args, parse_yml_or_exit, parse_config_or_exit
@@ -12,18 +13,41 @@ from .utils import get_sunrise_and_sunset, valid_dahua_sunrise_and_sunset
 
 def main():
     args = parse_args()
+
+    # Print version
     if args.version:
         print(__version__)
         return
+
+    # Print timezones
     if args.timezones:
         for t in pytz.all_timezones:
             print(t)
         return
 
     config = parse_config_or_exit(parse_yml_or_exit(args.path))
+    ret_code = 0
+
+    # Check all ipc
+    if args.check:
+        for c in config["ipc"]:
+            try:
+                ipc = get_ipc(c)
+                sunrise, sunset, switch_mode = ipc.get_sunrise_and_sunset(c["channel"])
+            except Exception as e:
+                print(traceback.format_exc())
+                logging.error(e)
+                ret_code = 1
+                continue
+            print(
+                f"{c['name']} sunrise is {sunrise}, sunset is {sunset}, and switch mode is set to {switch_mode}"
+            )
+        return ret_code
+
+    # Sync all ipc
     sunrise, sunset = get_sunrise_and_sunset(config["location"])
     print(
-        "Sunrise is at %s and sunset is at %s for %s"
+        "sunrise is %s and sunset is %s for %s"
         % (
             sunrise.strftime("%X"),
             sunset.strftime("%X"),
@@ -32,42 +56,44 @@ def main():
     )
     if not valid_dahua_sunrise_and_sunset(sunrise, sunset):
         logging.error(
-            "Daytime hours are not within a single day (e.g. sunrise 1:00 PM and sunset 12:01 AM the next day), check if your timezone and coordinates are correct"
+            "daytime hours are not within a single day (e.g. sunrise 1:00 PM and sunset 12:01 AM the next day), check if your timezone and coordinates are correct"
         )
         return 1
 
-    code = 0
-
     for c in config["ipc"]:
-        print("Syncing %s on channel %s..." % (c["name"], c["channel"]))
+        print("syncing %s on channel %s..." % (c["name"], c["channel"]))
 
         try:
-            if c["method"] == "cgi":
-                DahuaCgi(
-                    ip=c["ip"], username=c["username"], password=c["password"]
-                ).sync_sunrise_and_sunset(
-                    sunrise=sunrise,
-                    sunset=sunset,
-                    channel=c["channel"],
-                )
-            elif c["method"] == "rpc":
-                rpc = DahuaRpc(
-                    ip=c["ip"], username=c["username"], password=c["password"]
-                )
-                rpc.login()
-                rpc.sync_sunrise_and_sunset(
-                    sunrise=sunrise,
-                    sunset=sunset,
-                    channel=c["channel"],
-                )
-            else:
-                logging.error("Invalid method %s", c["method"])
-                continue
-        except Exception:
+            get_ipc(c).set_sunrise_and_sunset(
+                sunrise=sunrise,
+                sunset=sunset,
+                channel=c["channel"],
+            )
+        except Exception as e:
             print(traceback.format_exc())
-            code = 1
+            logging.error(e)
+            ret_code = 1
             continue
 
-        print("Synced %s on channel %s" % (c["name"], c["channel"]))
+        print("synced %s on channel %s" % (c["name"], c["channel"]))
 
-    return code
+    return ret_code
+
+
+def get_ipc(ipc_config: ConfigIPC) -> IPC:
+    if ipc_config["method"] == ConfigMethod.CGI:
+        return DahuaCgi(
+            ip=ipc_config["ip"],
+            username=ipc_config["username"],
+            password=ipc_config["password"],
+        )
+    if ipc_config["method"] == ConfigMethod.RPC:
+        rpc = DahuaRpc(
+            ip=ipc_config["ip"],
+            username=ipc_config["username"],
+            password=ipc_config["password"],
+        )
+        rpc.login()
+        return rpc
+
+    raise Exception("unknown method %s" % ipc_config["method"])