+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 
-- Option to interact with cameras via rpc instead of cgi.
+- Option to interact via rpc instead of cgi.
+- Add --check option to view ipc settings.
 
 ## [0.1.1] - 2022-06-16